sanitizers: ignore unknown addresses in reports
In some cases, the sanitizer logs contains addresses marked
as unknown. Here is an exampl ASan report due to an abort signal (ABRT):
 ==528025==ERROR: AddressSanitizer: ABRT on unknown address 0x03e800080e99 (pc ...

These addresses can be different for the same bug (PC and stacktrace are
the same) and as a result breaks error deduplication. Here, we ignore
addresses from the sanitizer reports if the address is markes as
unknown.